Add basic GitHub Actions log enhancements
Add a Python package that provides a ready interface to core GitHub
Actions workflow commands. Add some basic workflow log enhancements
which include grouping logs by repository and adding a notice for each
repository that has an event dispatched. This will make it easier to
see what the action is doing at a glance as well as make the logs more
easily navigated.