Check only most recent run on default branch
When the Markdown output is generated, the code previously determined
the failure status by checking if the most recent build passed or
failed. Often this build was a Dependabot or other PR build whose
failure should not determine the build status of the entire build. It
is better to instead check the last build that was built from the
default branch.

Now the Markdown output shows a failure for a project if and only if
the project is genuinely broken, i.e., if the default branch fails to
build.

See #68 for more details.