Make use of cisagov/action-job-preamble
This is preferable to rolling our own diagnostic job in this repo.

I also noticed that the APB rebuild process failed several times
recently because of the use of
GitHubSecurityLab/actions-permissions/monitor, which we have seen to
be problematic in the past.  Using cisagov/action-job-preamble cleanly
drops the use of this troublesome action.